chore: Move `EditHistory` to sidebar folder
